# packages/api-spec/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/menu/categories": {
      "get": {
        "operationId": "MenuController_listCategories",
        "summary": "List all categories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "menu"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "MenuController_createCategory",
        "summary": "Create a category",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created category",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          }
        },
        "tags": [
          "menu"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/menu/categories/{id}": {
      "get": {
        "operationId": "MenuController_getCategory",
        "summary": "Get category by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          }
        },
        "tags": [
          "menu"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "MenuController_updateCategory",
        "summary": "Update a category",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated category",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          }
        },
        "tags": [
          "menu"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/menu/items": {
      "get": {
        "operationId": "MenuController_listItems",
        "summary": "List all items, optionally filtered by category",
        "parameters": [
          {
            "name": "categoryId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "menu"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "MenuController_createItem",
        "summary": "Create an item",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          }
        },
        "tags": [
          "menu"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/menu/items/{id}": {
      "get": {
        "operationId": "MenuController_getItem",
        "summary": "Get item by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          }
        },
        "tags": [
          "menu"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "MenuController_updateItem",
        "summary": "Update an item",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          }
        },
        "tags": [
          "menu"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/day/open": {
      "post": {
        "operationId": "BusinessDayController_openDay",
        "summary": "Open a new business day",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenDayDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Business day opened",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayOpeningResponse"
                }
              }
            }
          },
          "409": {
            "description": "A business day is already open"
          }
        },
        "tags": [
          "day"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/day/close": {
      "post": {
        "operationId": "BusinessDayController_closeDay",
        "summary": "Close the current open business day",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseDayDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Business day closed with frozen totals",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloseDayResponse"
                }
              }
            }
          },
          "404": {
            "description": "No open business day to close"
          },
          "409": {
            "description": "Open/sent orders exist — cannot close"
          }
        },
        "tags": [
          "day"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/day/current": {
      "get": {
        "operationId": "BusinessDayController_getCurrent",
        "summary": "Get current open day with live X-report totals",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current open day or null"
          }
        },
        "tags": [
          "day"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/day": {
      "get": {
        "operationId": "BusinessDayController_list",
        "summary": "List past business days (paged)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paged list of business days"
          }
        },
        "tags": [
          "day"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/day/{id}": {
      "get": {
        "operationId": "BusinessDayController_getDay",
        "summary": "Get a business day by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Business day",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayOpeningResponse"
                }
              }
            }
          },
          "404": {
            "description": "Business day not found"
          }
        },
        "tags": [
          "day"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "SpicyHome POS API",
    "description": "REST API for the SpicyHome restaurant POS system",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Burgers"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32",
            "example": 0
          },
          "printerId": {
            "type": "integer",
            "format": "int64",
            "example": 1,
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "integer",
            "format": "int64",
            "example": 1700000000
          },
          "updatedAt": {
            "type": "integer",
            "format": "int64",
            "example": 1700000000
          },
          "createdBy": {
            "type": "integer",
            "format": "int64",
            "example": 1,
            "nullable": true
          },
          "updatedBy": {
            "type": "integer",
            "format": "int64",
            "example": 1,
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "sortOrder",
          "printerId",
          "isActive",
          "createdAt",
          "updatedAt",
          "createdBy",
          "updatedBy"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Burgers"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          "printerId": {
            "type": "integer",
            "format": "int64",
            "description": "Printer ID for kitchen routing"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Burgers"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          },
          "printerId": {
            "type": "integer",
            "format": "int64"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "ItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "categoryId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Zinger Burger"
          },
          "nameAr": {
            "type": "string",
            "example": "زنجر برجر",
            "nullable": true
          },
          "priceHalalas": {
            "type": "integer",
            "format": "int64",
            "example": 2300,
            "description": "VAT-inclusive price in halalas"
          },
          "vatRateBp": {
            "type": "integer",
            "format": "int32",
            "example": 1500,
            "description": "VAT rate in basis points"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32",
            "example": 0
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "integer",
            "format": "int64",
            "example": 1700000000
          },
          "updatedAt": {
            "type": "integer",
            "format": "int64",
            "example": 1700000000
          },
          "createdBy": {
            "type": "integer",
            "format": "int64",
            "example": 1,
            "nullable": true
          },
          "updatedBy": {
            "type": "integer",
            "format": "int64",
            "example": 1,
            "nullable": true
          }
        },
        "required": [
          "id",
          "categoryId",
          "name",
          "nameAr",
          "priceHalalas",
          "vatRateBp",
          "sortOrder",
          "isActive",
          "createdAt",
          "updatedAt",
          "createdBy",
          "updatedBy"
        ]
      },
      "CreateItemDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Zinger Burger"
          },
          "nameAr": {
            "type": "string",
            "example": "زنجر برجر"
          },
          "priceHalalas": {
            "type": "integer",
            "format": "int64",
            "example": 2300,
            "description": "VAT-inclusive price in halalas (23.00 SAR)"
          },
          "vatRateBp": {
            "type": "integer",
            "format": "int32",
            "default": 1500,
            "description": "VAT rate in basis points (1500 = 15%)"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "categoryId",
          "name",
          "priceHalalas"
        ]
      },
      "UpdateItemDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "nameAr": {
            "type": "string"
          },
          "priceHalalas": {
            "type": "integer",
            "format": "int64"
          },
          "vatRateBp": {
            "type": "integer",
            "format": "int32"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "OpenDayDto": {
        "type": "object",
        "properties": {
          "openingCashHalalas": {
            "type": "integer",
            "format": "int64",
            "description": "Opening cash counted in halalas",
            "example": 50000
          }
        },
        "required": [
          "openingCashHalalas"
        ]
      },
      "DayOpeningResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "businessDate": {
            "type": "string",
            "example": "2026-07-22"
          },
          "status": {
            "type": "string",
            "example": "open"
          },
          "openingCashHalalas": {
            "type": "integer",
            "format": "int64",
            "example": 50000
          },
          "openedAt": {
            "type": "integer",
            "format": "int64"
          },
          "openedBy": {
            "type": "integer",
            "format": "int64"
          },
          "closedAt": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "closedBy": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "closingCashHalalas": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "totalSalesHalalas": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "totalVatHalalas": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "orderCount": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "createdAt": {
            "type": "integer",
            "format": "int64"
          },
          "updatedAt": {
            "type": "integer",
            "format": "int64"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "updatedBy": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        },
        "required": [
          "id",
          "businessDate",
          "status",
          "openingCashHalalas",
          "openedAt",
          "openedBy",
          "createdAt",
          "updatedAt"
        ]
      },
      "CloseDayDto": {
        "type": "object",
        "properties": {
          "closingCashHalalas": {
            "type": "integer",
            "format": "int64",
            "description": "Closing cash counted in halalas",
            "example": 125000
          }
        },
        "required": [
          "closingCashHalalas"
        ]
      },
      "CloseDayResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "businessDate": {
            "type": "string",
            "example": "2026-07-22"
          },
          "status": {
            "type": "string",
            "example": "closed"
          },
          "openingCashHalalas": {
            "type": "integer",
            "format": "int64",
            "example": 50000
          },
          "openedAt": {
            "type": "integer",
            "format": "int64"
          },
          "openedBy": {
            "type": "integer",
            "format": "int64"
          },
          "closedAt": {
            "type": "integer",
            "format": "int64"
          },
          "closedBy": {
            "type": "integer",
            "format": "int64"
          },
          "closingCashHalalas": {
            "type": "integer",
            "format": "int64",
            "example": 125000
          },
          "totalSalesHalalas": {
            "type": "integer",
            "format": "int64",
            "example": 46000
          },
          "totalVatHalalas": {
            "type": "integer",
            "format": "int64",
            "example": 6000
          },
          "orderCount": {
            "type": "integer",
            "format": "int64",
            "example": 5
          }
        },
        "required": [
          "id",
          "businessDate",
          "status",
          "openingCashHalalas",
          "openedAt",
          "openedBy",
          "closedAt",
          "closedBy",
          "closingCashHalalas",
          "totalSalesHalalas",
          "totalVatHalalas",
          "orderCount"
        ]
      }
    }
  }
}